+++ threedi_verification/models.py
@@ -13,15 +13,27 @@ import jsonfield
 
 logger = logging.getLogger(__name__)
 
+# four characters max
+SUBGRID = 'SUBG'
+FLOW = 'FLOW'
+LIBRARIES = (
+    (SUBGRID, 'Subgrid'),
+    (FLOW, 'Flow'),
+)
+
 
 class TestCase(models.Model):
 
-    filename = models.CharField(
-        verbose_name=_("filename"),
+    path = models.CharField(
+        verbose_name=_("path"),
         unique=True,
-        help_text=_(
-            "MDU filename, including path, inside the 'testbank' directory"),
+        help_text=_("Path to model input (mdu or other input, e.g., a model "
+                    "dir), w.r.t. the 'testbank' (TESTCASES_ROOT) directory"),
         max_length=255)
+    library = models.CharField(
+        max_length=4,
+        choices=LIBRARIES,
+        default=SUBGRID)
     info = models.TextField(
         verbose_name=_("information"),
         blank=True,
@@ -36,10 +48,10 @@ class TestCase(models.Model):
     class Meta:
         verbose_name = _("test case")
         verbose_name_plural = _("test cases")
-        ordering = ['filename']
+        ordering = ['path']
 
     def __unicode__(self):
-        return _("test case %s") % self.filename
+        return _("test case %s") % self.path
 
     def get_absolute_url(self):
         return reverse('threedi_verification.test_case',
@@ -47,8 +59,9 @@ class TestCase(models.Model):
 
     @cached_property
     def pretty_name(self):
-        name = self.filename.split('/')[-1]
-        name = name.rstrip('.mdu')
+        name = self.path.split('/')[-1]
+        if self.library == SUBGRID:
+            name = name.rstrip('.mdu')
         if self.info:
             first_line = self.info.split('\n')[0].strip()
             return "%s (%s)" % (first_line, name)
@@ -61,6 +74,10 @@ class TestCase(models.Model):
             test_case_version__test_case=self)
         if test_runs.exists():
             return test_runs.first()
+
+    @cached_property
+    def library_name(self):
+        return dict(LIBRARIES).get(self.library)
 
 
 class TestCaseVersion(models.Model):
@@ -89,20 +106,25 @@ class TestCaseVersion(models.Model):
 
 class LibraryVersion(models.Model):
 
+    library = models.CharField(
+        max_length=4,
+        choices=LIBRARIES,
+        default=SUBGRID)
     last_modified = models.DateTimeField(
         unique=True,
         verbose_name=_("last modified"))
     num_test_cases = models.IntegerField(
-        default = 0,
+        default=0,
         verbose_name=_("number of test cases when library was first found"))
 
     class Meta:
-        verbose_name = _("library")
-        verbose_name_plural = _("libraries")
+        verbose_name = _("library version")
+        verbose_name_plural = _("library versions")
         ordering = ['-last_modified']
 
     def __unicode__(self):
-        return _("library version of %s") % self.last_modified
+        return _("%s library version of %s") % (self.library_name,
+                                                self.last_modified)
 
     def get_absolute_url(self):
         return reverse('threedi_verification.library_version',
@@ -123,7 +145,11 @@ class LibraryVersion(models.Model):
         # This one probably needs caching.
         return range(len([test_run for test_run in self.test_runs.filter(
             test_case_version__test_case__has_csv=True)
-                          if test_run.has_crashed]))
+            if test_run.has_crashed]))
+
+    @cached_property
+    def library_name(self):
+        return dict(LIBRARIES).get(self.library)
 
 
 class TestRun(models.Model):
@@ -163,7 +189,7 @@ class TestRun(models.Model):
 
     @cached_property
     def num_wrong(self):
-        if not 'instruction_reports' in self.report:
+        if 'instruction_reports' not in self.report:
             return 0
         return len(
             [ir for ir in self.report['instruction_reports']
@@ -171,7 +197,7 @@ class TestRun(models.Model):
 
     @cached_property
     def num_right(self):
-        if not 'instruction_reports' in self.report:
+        if 'instruction_reports' not in self.report:
             return 0
         return len(
             [ir for ir in self.report['instruction_reports']